+++ Runtime/INavTransition.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using Cysharp.Threading.Tasks;
 
-namespace KidzDev.Unity.SceneNavigator
+namespace KidzDev.Unity.ScreenNavigator
 {
     /// <summary>
     /// The animation seam. Animates the visual move from <paramref name="from"/> to <paramref name="to"/>.