+++ Runtime/Providers/RegistryScreenProvider.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using Cysharp.Threading.Tasks;
 
-namespace KidzDev.Unity.SceneNavigator
+namespace KidzDev.Unity.ScreenNavigator
 {
     /// <summary>
     /// The simplest provider: a dictionary of pre-built, scene-resident screens registered by key.