+++ Runtime/Transitions/InstantTransition.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using Cysharp.Threading.Tasks;
 
-namespace KidzDev.Unity.SceneNavigator
+namespace KidzDev.Unity.ScreenNavigator
 {
     /// <summary>
     /// The default transition: no animation. Completes synchronously. The navigator handles activation,